Extract shard transfer implementations into modules
